+++ src/main/java/com/seoulotakus/takumapbe/common/auth/dto/request/IdCheckRequestDTO.java
@@ -1,0 +1,15 @@
+package com.seoulotakus.takumapbe.common.auth.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class IdCheckRequestDTO {
+
+    @NotBlank
+    private String userId;
+}
