+++ src/main/java/com/seoulotakus/takumapbe/common/auth/dto/request/LoginRequestDTO.java
@@ -1,0 +1,18 @@
+package com.seoulotakus.takumapbe.common.auth.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class LoginRequestDTO {
+
+    @NotBlank
+    private String userId;
+
+    @NotBlank
+    private String password;
+}
