+++ src/main/java/com/seoulotakus/takumapbe/common/auth/dto/response/AccessTokenResponseDTO.java
@@ -1,0 +1,13 @@
+package com.seoulotakus.takumapbe.common.auth.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class AccessTokenResponseDTO {
+
+    private String accessToken;
+}
