+++ src/main/java/com/seoulotakus/takumapbe/common/config/PrincipalDetails.java
@@ -1,0 +1,75 @@
+package com.seoulotakus.takumapbe.common.config;
+
+import com.seoulotakus.takumapbe.domain.user.entity.UserEntity;
+import lombok.Data;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+
+/** 다시 작성해야함 **/
+@Data
+public class PrincipalDetails implements UserDetails, OAuth2User {
+    
+    private UserEntity user;  // 콤포지션
+    private Map<String, Object> attributes;
+
+    // 자체 로그인 생성자
+    public PrincipalDetails(UserEntity user){
+        this.user = user;
+    }
+
+    // 소셜 로그인 생성자(OAuth)
+    public PrincipalDetails(UserEntity user, Map<String, Object> attributes){
+        this.user = user;
+        this.attributes = attributes;
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return attributes;
+    }
+
+    @Override
+    public String getName() {
+        return user.getNickname();
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of();
+    }
+
+    @Override
+    public String getPassword() {
+        return user.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getNickname();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return UserDetails.super.isAccountNonExpired();
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return UserDetails.super.isAccountNonLocked();
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return UserDetails.super.isCredentialsNonExpired();
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return UserDetails.super.isEnabled();
+    }
+}
