--- src/main/java/com/seoulotakus/takumapbe/domain/auth/dto/request/LoginRequestDTO.java
@@ -1,18 +1,0 @@
-package com.seoulotakus.takumapbe.domain.auth.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class LoginRequestDTO {
-
-    @NotBlank
-    private String userId;
-
-    @NotBlank
-    private String password;
-}
