--- src/main/java/com/seoulotakus/takumapbe/domain/auth/dto/request/SignUpRequestDTO.java
@@ -1,34 +1,0 @@
-package com.seoulotakus.takumapbe.domain.auth.dto.request;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignUpRequestDTO {
-
-    @NotBlank
-    private String userId;
-
-    @NotBlank
-    // 영문자(대소문자), 숫자, 특수문자(!@#$%^&+=) 포함 8~15글자
-    @Pattern(regexp = "^.*(?=^.{8,15}$)(?=.*\\d)(?=.*[a-zA-Z])(?=.*[!@#$%^&+=]).*$")
-    private String password;
-
-    @Email
-    @NotBlank
-    private String email;
-
-    @NotBlank
-    private String nickname;
-
-    @NotBlank
-    private String certificationNumber;
-}
