--- src/main/java/com/seoulotakus/takumapbe/domain/auth/provider/jwt/JwtProvider.java
@@ -1,60 +1,0 @@
-package com.seoulotakus.takumapbe.domain.auth.provider.jwt;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwt;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
-import java.util.Date;
-
-@Component
-public class JwtProvider {
-
-    @Value("${jwt.secret}")
-    private String secretKey;
-
-    // JWT 만드는 메소드
-    public String create(String userId){
-
-        // 유효기간 - 1시간
-        Date expiredDate = Date.from(Instant.now().plus(1, ChronoUnit.HOURS));
-        // 시크릿 키
-        Key key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-        // JWT
-        String jwt = Jwts.builder()
-                .signWith(key, SignatureAlgorithm.HS256)
-                .setSubject(userId).setIssuedAt(new Date()).setExpiration(expiredDate)
-                .compact();
-
-        return jwt;
-    }
-
-    // jwt 검증하는 메소드
-    public String validate(String jwt){
-        String subject = null;
-        Key key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-
-        try{
-            Claims claims = Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(jwt)
-                    .getBody();
-
-            subject = claims.getSubject();
-
-        }catch (Exception exception){
-            exception.printStackTrace();
-            return null;
-        }
-
-        return subject;
-    }
-}
