+++ src/main/java/com/seoulotakus/takumapbe/global/config/WebConfig.java
@@ -15,6 +15,7 @@ public class WebConfig implements WebMvcConfigurer {
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**")
                 .allowedOriginPatterns(
+                        "http://localhost:3000",
                         "https://takumap.vercel.app",
                         "https://*.vercel.app")
                 .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH")
