--- src/main/java/com/seoulotakus/takumapbe/global/config/oauth/dto/OAuthAttributes.java
@@ -1,96 +1,0 @@
-package com.seoulotakus.takumapbe.global.config.oauth.dto;
-
-import com.seoulotakus.takumapbe.domain.user.entity.UserEntity;
-import com.seoulotakus.takumapbe.domain.user.enums.Provider;
-import com.seoulotakus.takumapbe.domain.user.enums.UserRole;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.Map;
-
-@Getter
-public class OAuthAttributes {
-
-    private Map<String, Object> attributes;
-    private String nameAttributeKey;
-    private String name;
-    private String email;
-    private Provider provider;
-    private String providerId;
-
-
-    @Builder
-    public OAuthAttributes(Map<String, Object> attributes, String nameAttributeKey, String name,
-                           String email, Provider provider, String providerId) {
-        this.attributes = attributes;
-        this.nameAttributeKey = nameAttributeKey;
-        this.name = name;
-        this.email = email;
-        this.provider = provider;
-        this.providerId = providerId;
-    }
-
-    // OAuth2User에서 반환하는 사용자 정보는 Map이기 때문에 값 하나하나를 변환해야 함
-    public static OAuthAttributes of(String registrationId, String userNameAttributeName, Map<String,
-            Object> attributes) {
-        if ("naver".equals(registrationId)) {
-            return ofNaver("id", attributes);
-        }
-        if ("kakao".equals(registrationId)) {
-            return ofKakao("id", attributes);
-        }
-        return ofGoogle(userNameAttributeName, attributes);
-    }
-
-    private static OAuthAttributes ofGoogle(String userNameAttributeName, Map<String, Object>
-            attributes) {
-        return OAuthAttributes.builder()
-                .name((String) attributes.get("name"))
-                .email((String) attributes.get("email"))
-                .provider(Provider.GOOGLE)
-                .providerId((String) attributes.get("sub"))
-                .attributes(attributes)
-                .nameAttributeKey(userNameAttributeName)
-                .build();
-    }
-
-    private static OAuthAttributes ofNaver(String userNameAttributeName, Map<String, Object>
-            attributes) {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-        return OAuthAttributes.builder()
-                .name((String) response.get("name"))
-                .email((String) response.get("email"))
-                .provider(Provider.NAVER)
-                .providerId((String) response.get("id"))
-                .attributes(response)
-                .nameAttributeKey(userNameAttributeName)
-                .build();
-    }
-
-    private static OAuthAttributes ofKakao(String userNameAttributeName, Map<String, Object>
-            attributes) {
-        Map<String, Object> kakaoAccount = (Map<String, Object>) attributes.get("kakao_account");
-        Map<String, Object> profile = (Map<String, Object>) kakaoAccount.get("profile");
-
-        return OAuthAttributes.builder()
-                .name((String) profile.get("nickname"))
-                .email((String) kakaoAccount.get("email"))
-                .provider(Provider.KAKAO)
-                .providerId(String.valueOf(attributes.get("id")))
-                .attributes(attributes)
-                .nameAttributeKey(userNameAttributeName)
-                .build();
-    }
-
-
-    // User 엔티티 생성
-    public UserEntity toEntity() {
-        return UserEntity.builder()
-                .nickname(name)
-                .email(email)
-                .userRole(UserRole.ROLE_USER) // 기본 권한
-                .provider(provider)
-                .providerId(providerId)
-                .build();
-    }
-}
