+++ src/main/java/com/seoulotakus/takumapbe/global/exception/BusinessException.java
@@ -1,13 +1,16 @@
 package com.seoulotakus.takumapbe.global.exception;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 
 @Getter
-@RequiredArgsConstructor
 public class BusinessException extends RuntimeException {
 
     private final ErrorCode errorCode;
+
+    public BusinessException(ErrorCode errorCode){
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
 
     public BusinessException(ErrorCode errorCode, Throwable cause){
         super(errorCode.getMessage(), cause);
